Fixed codestyle issues and renamed files
